feat: add LiteLLMAPIBackend and DeprecBackend ; changed structure of the project ; with bus
